+++ pulse_lib/segments/segment_pulse.py
@@ -1,15 +1,15 @@
-"""
-Class that is used to make DC pulses.
-"""
+from dataclasses import dataclass
+from typing import Callable
+
 import numpy as np
 
 from pulse_lib.configuration.iq_channels import IQ_out_channel_info
-from pulse_lib.segments.segment_base import segment_base
-from pulse_lib.segments.utility.data_handling_functions import loop_controller
 from pulse_lib.segments.data_classes.data_pulse import pulse_data, custom_pulse_element, pulse_delta
 from pulse_lib.segments.data_classes.data_IQ import IQ_data_single
+from pulse_lib.segments.segment_base import segment_base
 from pulse_lib.segments.segment_IQ import segment_IQ
-from dataclasses import dataclass
+from pulse_lib.segments.utility.data_handling_functions import loop_controller
+
 
 @dataclass
 class IQ_render_info:
@@ -19,10 +19,12 @@ class IQ_render_info:
     virtual_channel: segment_IQ
     out_channel: IQ_out_channel_info
 
+
 class segment_pulse(segment_base):
     '''
     Class defining single segments for one sequence.
     '''
+
     def __init__(self, name, segment_type='render', hres=False):
         '''
         Args:
@@ -65,17 +67,6 @@ class segment_pulse(segment_base):
             amplitude : total hight of the ramp, starting from the base point
             keep_amplitude : when pulse is done, keep reached amplitude for time infinity
         '''
-        return self._add_ramp(start, stop, start_amplitude, stop_amplitude, keep_amplitude)
-
-    def _add_ramp(self, start, stop, start_amplitude, stop_amplitude, keep_amplitude=False):
-        '''
-        Makes a linear ramp (with start and stop amplitude)
-        Args:
-            start (double) : starting time of the ramp
-            stop (double) : stop time of the ramp
-            amplitude : total hight of the ramp, starting from the base point
-            keep_amplitude : when pulse is done, keep reached amplitude for time infinity
-        '''
         if start != stop:
             ramp = (stop_amplitude-start_amplitude) / (stop-start)
             self.data_tmp.add_delta(pulse_delta(start + self.data_tmp.start_time,
@@ -103,7 +94,7 @@ class segment_pulse(segment_base):
     @loop_controller
     def wait(self, wait):
         '''
-        wait for x ns after the lastest wave element that was programmed.
+        wait for x ns after the lastest wave element of the segment.
         Args:
             wait (double) : time in ns to wait
         '''
@@ -127,13 +118,50 @@ class segment_pulse(segment_base):
                                                  stop + self.data_tmp.start_time,
                                                  amp, freq,
                                                  phase_offset,
-                                                 None, # no envelope
+                                                 None,  # no envelope
                                                  self.name,
                                                  coherent_pulsing=False))
         return self.data_tmp
 
     @loop_controller
-    def add_custom_pulse(self, start, stop, amplitude, custom_func, **kwargs):
+    def add_custom_pulse_v2(self,
+                            start: float,
+                            stop: float,
+                            amplitude: float,
+                            custom_func_v2: Callable[[np.ndarray, float, float, ...], np.ndarray],
+                            **kwargs):
+        """
+        Adds a custom pulse to this segment.
+        Args:
+            start (double) : start time in ns of the pulse
+            stop (double) : stop time in ns of the pulse
+            amplitude (double) : amplitude of the pulse
+            custom_func_v2: function to generate the samples for this pulse. It must return a 1D numpy array.
+            kwargs: keyword arguments passed into the custom_func
+
+        Example:
+            def hamming_pulse(t: np.ndarray, duration: float, amplitude: float, alpha: float):
+                y = np.ones(t.shape)*alpha
+                # Note: t[0] is <= 0.0
+                y[0] = 2*alpha-1
+                y[-1] = 2*alpha-1
+                y[1:-1] += (alpha-1) * np.cos(2*np.pi*t[1:-1]/(duration-(t[1]-t[0])))
+                return y * amplitude
+
+            seg.P1.add_custom_pulse_v2(0, 10, 142.0, hamming_pulse, alpha=0.54)
+        """
+        pulse_data = custom_pulse_element(start + self.data_tmp.start_time, stop + self.data_tmp.start_time,
+                                          amplitude, func_v2=custom_func_v2, kwargs=kwargs)
+        self.data_tmp.add_custom_pulse_data(pulse_data)
+        return self.data_tmp
+
+    @loop_controller
+    def add_custom_pulse(self,
+                         start: float,
+                         stop: float,
+                         amplitude: float,
+                         custom_func: Callable[[float, float, float, ...], np.ndarray],
+                         **kwargs):
         """
         Adds a custom pulse to this segment.
         Args:
@@ -149,10 +177,9 @@ class segment_pulse(segment_base):
                 return signal.windows.tukey(n_points, alpha) * amplitude
 
             seg.P1.add_custom_pulse(0, 10, 142.0, tukey_pulse, alpha=0.5)
-
         """
         pulse_data = custom_pulse_element(start + self.data_tmp.start_time, stop + self.data_tmp.start_time,
-                                          amplitude, custom_func, kwargs)
+                                          amplitude, func=custom_func, kwargs=kwargs)
         self.data_tmp.add_custom_pulse_data(pulse_data)
         return self.data_tmp